independent select components, reorganize cli guide
