+++ docs/.vitepress/components.d.ts
@@ -11,6 +11,7 @@ declare module '@vue/runtime-core' {
     BlockQuote: typeof import('./components/BlockQuote.vue')['default']
     Contributors: typeof import('./components/Contributors.vue')['default']
     DatabaseBlock: typeof import('./components/DatabaseBlock.vue')['default']
+    DatabaseSelect: typeof import('./components/DatabaseSelect.vue')['default']
     ElCheckbox: typeof import('element-plus/es')['ElCheckbox']
     ElInput: typeof import('element-plus/es')['ElInput']
     ElOption: typeof import('element-plus/es')['ElOption']
@@ -19,6 +20,7 @@ declare module '@vue/runtime-core' {
     ElSelect: typeof import('element-plus/es')['ElSelect']
     FeaturesList: typeof import('./components/FeaturesList.vue')['default']
     LanguageBlock: typeof import('./components/LanguageBlock.vue')['default']
+    LanguageSelect: typeof import('./components/LanguageSelect.vue')['default']
     ListItem: typeof import('./components/ListItem.vue')['default']
     Logo: typeof import('./components/Logo.vue')['default']
     Select: typeof import('./components/Select.vue')['default']
