+++ docs/guides/cli/configuration.md
@@ -8,15 +8,3 @@ A generated application comes with a schema that validates the initial configura
 
 The configuration [schema definition](../../api/schema/index.md) can be found in `configuration.ts`. It is used as a [configuration schema](../../api/configuration.md#configuration-validation) and loads some default schemas for authentication and database connection configuration and adds values for `host`, `port` and the `public` hosted file folder. The types of this schema are also used for `app.get()` and `app.set()` [typings](./typescript.md). The initial configuration schema will be validated on application startup when calling [`app.listen()`](../../api/application.md#listenport) or [`app.setup()`](../../api/application.md#setupserver).
 
-## Folders
-
-The source and test folders to which files are generated is set in the `package.json`. To change them, rename the `src/` or `test/` folder to what you want it to and then update `package.json` `directories` section accordingly:
-
-```json
-{
-  "directories": {
-    "lib": "api/src",
-    "test": "api/test"
-  }
-}
-```
